Se mueven los archivos de Gabriela a la carpeta correcta para que funcione
