+++ build.gradle
@@ -4,11 +4,9 @@ buildscript {
     repositories {
         google()
         jcenter()
-        
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.2'
-        
+        classpath 'com.android.tools.build:gradle:3.6.3'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -18,7 +16,6 @@ allprojects {
     repositories {
         google()
         jcenter()
-        
     }
 }
